Consolidated and slightly improved FileGetContents and CURLClient tests.
